feat: fix of all project warnings
