+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ccmobile",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "prestart": "sb-rn-get-stories",
