+++ src/bll/store/store.ts
@@ -4,16 +4,6 @@ import {quizReducer} from '../quizReducer';
 import {processReducer} from '@src/bll/processReducer';
 import {resultReducer} from '@src/bll/resultReducer';
 import {authReducer} from '@src/bll/authReducer';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import {
-  persistReducer,
-  FLUSH,
-  REHYDRATE,
-  PAUSE,
-  PERSIST,
-  PURGE,
-  REGISTER,
-} from 'redux-persist';
 
 const rootReducer = combineReducers({
   app: appReducer,
@@ -22,19 +12,9 @@ const rootReducer = combineReducers({
   processReducer,
   resultReducer,
 });
-const persistConfig = {
-  key: 'root',
-  storage: AsyncStorage,
-};
-const persistedReducer = persistReducer(persistConfig, rootReducer);
+
 export const store = configureStore({
-  reducer: persistedReducer,
-  middleware: getDefaultMiddleware =>
-    getDefaultMiddleware({
-      serializableCheck: {
-        ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER],
-      },
-    }),
+  reducer: rootReducer,
 });
 export type RootStateT = ReturnType<typeof store.getState>;
 export type AppDispatchT = typeof store.dispatch;
